subtask1/requirements.txt
tensorflow-gpu==1.6.0